+++ packages/rrweb-snapshot/src/snapshot.ts
@@ -19,10 +19,9 @@ import {
   isShadowRoot,
   maskInputValue,
   isNativeShadowDom,
-  getCssRulesString,
+  stringifyStylesheet,
   getInputType,
   toLowerCase,
-  validateStringifiedCssRule,
 } from './utils';
 
 let _id = 1;
@@ -49,20 +48,6 @@ function getValidTagName(element: HTMLElement): Lowercase<string> {
   }
 
   return processedTagName;
-}
-
-function stringifyStyleSheet(sheet: CSSStyleSheet): string {
-  if (sheet.cssRules.length === 0) {
-    return '';
-  }
-
-  const buffer: string[] = new Array<string>(sheet.cssRules.length);
-
-  for (let i = 0; i < sheet.cssRules.length; i++) {
-    const rule = sheet.cssRules[i];
-    buffer[i] = rule.cssText ? validateStringifiedCssRule(rule.cssText) : '';
-  }
-  return buffer.join('');
 }
 
 function extractOrigin(url: string): string {
@@ -576,7 +561,7 @@ function serializeTextNode(
         // to _only_ include the current rule(s) added by the text node.
         // So we'll be conservative and keep textContent as-is.
       } else if ((n.parentNode as HTMLStyleElement).sheet?.cssRules) {
-        textContent = stringifyStyleSheet(
+        textContent = stringifyStylesheet(
           (n.parentNode as HTMLStyleElement).sheet!,
         );
       }
@@ -673,7 +658,7 @@ function serializeElementNode(
     });
     let cssText: string | null = null;
     if (stylesheet) {
-      cssText = getCssRulesString(stylesheet);
+      cssText = stringifyStylesheet(stylesheet);
     }
     if (cssText) {
       delete attributes.rel;
@@ -688,7 +673,7 @@ function serializeElementNode(
     // TODO: Currently we only try to get dynamic stylesheet when it is an empty style element
     !(n.innerText || n.textContent || '').trim().length
   ) {
-    const cssText = getCssRulesString(
+    const cssText = stringifyStylesheet(
       (n as HTMLStyleElement).sheet as CSSStyleSheet,
     );
     if (cssText) {